requirements.txt
# Основные библиотеки для ML
xgboost>=1.6.0
scikit-learn>=1.1.0
pandas>=1.5.0
numpy>=1.21.0

# Для работы с данными
yfinance==0.2.28
ccxt>=2.0.0
requests>=2.28.0

# Для визуализации
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Для анализа временных рядов
ta>=0.10.0
statsmodels==0.14.1

# Для конфигурации
python-dotenv==1.0.0
pyyaml>=6.0

# Для логирования
loguru>=0.6.0

# Для Jupyter notebooks
jupyter==1.0.0
ipykernel==6.27.1

# Для тестирования
pytest==7.4.3
pytest-cov==4.1.0

# Для балансировки данных
imbalanced-learn>=0.10.0

# Для Bayesian optimization
scikit-optimize==0.9.0
joblib==1.3.2

# Для парсинга новостей
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Для работы с данными
python-dateutil>=2.8.0

# Для оптимизации гиперпараметров
optuna>=3.0.0 